Added availabilities to offers
Added availabilities to offers so that user's may see
a list of times in which an offer is available to make contracts
on. Added contraints to contract creation that a contract must 1:
have an offer_uuid that links to an existing offer and 2: start/
end_times for a contract must fall within an offer's availability range.
Added filtering to grab offers by their availabilities.