Rework destroy to cancel
Changed the way the DELETE requets on offers and contracts
function. Instead of deleting the object from the database,
we instead change the object's status to a new 'cancelled'
status. Additionally, cancelling an offer will cancel all
associated contracts and free up the availability on the offer.
Moved some api logic to the objects layer as well for consistency.