Locks on objects
Adding syncrhonization using the oslo concurrency
lockutils module. This is needed to address race conditions
when creating contracts and offers. This commit prevents
conflicting contracts from being created on an offer,
conflicting offers being created on a resource, and a contract
from being created on an offer being deleted. Also refactors
the offer and contract controllers uuid and name resolution.