Ensured exception handling logs exception type
If an operation on an object fails with an exception, the exception
message will be displayed, but currently, the type is excluded. This is
not ideal-- if an exception like OutOfMemory('64G/64G') was raised, the
message may not provide enough context to solve the root problem. This
patch addresses this.

Addresses #267.