Add integration tests for container image
Verify that we can successfully build an esi-leap container image and that
we can successfully start a container from the image.

These tests are run automatically by .github/workflows/build-image.yaml,
which on pushes to the `master` branch will also push the image into the
ghcr.io container repository.

The integration tests introduced in this commit are minimal in nature and
are ripe for further work if someone is feeling inspired.

To run the integration tests manually:

    tox -e integration