Fix test warnings and errors under pytest
I tried running the test suite using pytest, and while it mostly works,
there were a number of errors and warnings that cropped up. Primarily:

1. We need to catch exceptions in threads spawned during the test process.

2. We need to pass valid UUIDs to anything expecting a UUID; failure to do
    so yields a warning. So many warnings.

With this commit, the tests -- with the exception of
TestRequestContextSerializer -- run cleanly under pytest. There are still
warnings, but they all come from third party modules and so are somebody
else's problem.

The issues with TestRequestContextSerializer are resolved in the
following commit.